+++ lib/builders/view_model_builder.dart
@@ -2,8 +2,10 @@ part of mvvm;
 
 class ViewModelBuilder<TViewModel extends ViewModel>
     extends ModelBuilder<TViewModel, BaseEvent> {
+  final FutureOr<void> Function()? onLoad;
   const ViewModelBuilder({
     required TViewModel viewModel,
+    this.onLoad,
     required Widget Function(BuildContext context, TViewModel viewModel,
             BaseEvent? previousEvent, BaseEvent? event)
         builder,
@@ -23,16 +25,33 @@ class ViewModelBuilder<TViewModel extends ViewModel>
 
 class ViewModelBuilderState<TViewModel extends ViewModel>
     extends State<ViewModelBuilder<TViewModel>> {
-  late BaseEvent? oldEvent;
-  late BaseEvent? event;
+  BaseEvent? oldEvent;
+  BaseEvent? event;
   late StreamSubscription<BaseEvent> subscription;
+  bool loaded = false;
 
   @override
   void initState() {
-    setState(() {
-      oldEvent = null;
-      event = null;
-    });
+    super.initState();
+
+    if (widget.onLoad == null) {
+      loaded = true;
+      createSubscription();
+    } else {
+      WidgetsBinding.instance!.addPostFrameCallback((_) async {
+        await widget.onLoad!();
+
+        setState(
+          () {
+            loaded = true;
+            createSubscription();
+          },
+        );
+      });
+    }
+  }
+
+  void createSubscription() {
     subscription = widget.model.changes.listen((newEvent) {
       if (newEvent == oldEvent) return;
 
@@ -41,8 +60,6 @@ class ViewModelBuilderState<TViewModel extends ViewModel>
         event = newEvent;
       });
     });
-
-    super.initState();
   }
 
   @override